Enemy Ship Fix and Death for Player
In this commit I fixed how the enemy ship looks at the player in the update void section. Then added a collision if statement that would kill the asteroid when it collides with a bullet. With also fixing the spawning for the asteroid so there can be no more than three at a time.